# public/swagger/collection/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Negotium Solutions Swagger API Documentation",
        "description": "This is a Negotium swagger API documentation. You can find out more about Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net,#swagger](http://swagger.io/irc/).For this sample, you can use the api key `special-key` to test the authorization filters",
        "contact": {
            "name": "Admin",
            "email": "admin@negotium-solutions.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://api.negotium-solutions.com/api",
            "description": "This is the Negotium Solutions Swagger API Server"
        }
    ],
    "paths": {
...
        "/{tenant}/step/{id}": {
            "get": {
                "tags": [
                    "Step"
                ],
                "summary": "Get a Step",
                "description": "Authenticate using a bearer token",
                "operationId": "getStep",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Step Id",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/{tenant}/step": {
            "get": {
                "tags": [
                    "Step"
                ],
                "summary": "Get Steps",
                "description": "Authenticate using a bearer token",
                "operationId": "getSteps",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/{tenant}/step/create": {
            "post": {
                "tags": [
                    "Step"
                ],
                "summary": "Create a new step",
                "description": "Authenticate using a bearer token",
                "operationId": "createStep",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Input validation error"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/{tenant}/step/update/{id}": {
            "put": {
                "tags": [
                    "Step"
                ],
                "summary": "Update aStep",
                "description": "Authenticate using a bearer token",
                "operationId": "updateStep",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Step Id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Input validation error"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/{tenant}/step/delete/{id}": {
            "delete": {
                "tags": [
                    "Step"
                ],
                "summary": "Delete a Step by ID.",
                "description": "Authenticate using a bearer token",
                "operationId": "deleteStepById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Not found"
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
...
}